+++ Jeu/src/moteur/MoteurJeu.java
@@ -53,9 +53,15 @@ public class MoteurJeu extends Application {
     public static Stage primaryStage;
     ModeleLabyrinth laby = (ModeleLabyrinth) MoteurJeu.jeu;
     // initialisation du canvas de dessin et du container
-    final Canvas canvas = new Canvas();
+    Canvas canvas = new Canvas();
     final Pane canvasContainer = new Pane(canvas);
     private Label objectifLabel;
+    ConfigWindow configWindow = new ConfigWindow(laby,primaryStage,canvas,this);
+    ViewGraphiqueDirect viewGraphiqueDirect ;
+    ViewGraphiqueObjectif viewGraphiqueObjectif;
+    ViewLabyrinth viewLabyrinth;
+    ViewLogs viewLogs ;
+    ViewGraphique viewGraphique;
 
     /**
      * lancement d'un jeu
@@ -84,298 +90,11 @@ public class MoteurJeu extends Application {
      * creation de l'application avec juste un canvas et des statistiques
      */
     public void start(Stage primaryStage) throws IOException {
-        MoteurJeu.primaryStage = primaryStage;
-        // Création des vues
-        ViewLabyrinth viewLabyrinth = new ViewLabyrinth(laby, canvas);
-        laby.registerObserver(viewLabyrinth);
-
-        // Configuration de la fenêtre de dialogue
-        Stage dialogStage = new Stage();
-        dialogStage.setTitle("Configuration du Labyrinthe");
-
-        // Conteneur principal avec style moderne
-        VBox root = new VBox(20);
-        root.setPadding(new Insets(30));
-        root.setStyle("-fx-background-color: #f5f5f5; "
-                + "-fx-background-radius: 10; "
-                + "-fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.1), 10, 0, 0, 0);");
-
-        // En-tête
-        Label titleLabel = new Label("Configuration de la partie");
-        titleLabel.setStyle("-fx-font-size: 24px; "
-                + "-fx-font-weight: bold; "
-                + "-fx-text-fill: #2c3e50; "
-                + "-fx-padding: 0 0 20 0;");
-        titleLabel.setAlignment(Pos.CENTER);
-        titleLabel.setMaxWidth(Double.MAX_VALUE);
-
-        // Section de sélection du labyrinthe préfait
-        ComboBox<String> configPrefaiteComboBox = new ComboBox<>();
-        configPrefaiteComboBox.getItems().addAll(
-                "Aucune",
-                "Chemin safe",
-                "Aucun chemin safe",
-                "Une tour unique",
-                "Nombreuses tours centrales",
-                "Tours inaccessibles"
-        );
-        configPrefaiteComboBox.setValue("Aucune");
-        configPrefaiteComboBox.setId("configPrefaiteComboBox");
-        configPrefaiteComboBox.setStyle("-fx-background-color: #f8f9fa; "
-                + "-fx-background-radius: 5; "
-                + "-fx-border-color: #dee2e6; "
-                + "-fx-border-radius: 5; "
-                + "-fx-padding: 5; ");
-
-        VBox configPrefaiteSection = createSection("Choisir une configuration de test", configPrefaiteComboBox);
-
-        // Map des labyrinthes
-        Map<String, String> labyrinthMap = new HashMap<>();
-        labyrinthMap.put("Petit", "Ressources/Labyrinthe1.txt");
-        labyrinthMap.put("Grand", "Ressources/Labyrinthe2.txt");
-        labyrinthMap.put("Large", "Ressources/Labyrinthe3.txt");
-        labyrinthMap.put("Large sans safe", "Ressources/Labyrinthe3_bis.txt");
-        labyrinthMap.put("Ligne", "Ressources/Ligne.txt");
-        labyrinthMap.put("Nombreuses tours centrales", "Ressources/LignePlusieursTours.txt");
-        labyrinthMap.put("Tours inaccessibles", "Ressources/ToursInaccessibles.txt");
-
-        // Section de sélection du labyrinthe
-        ComboBox<String> labyrinthComboBox = new ComboBox<>();
-        labyrinthComboBox.getItems().addAll(
-                "Petit",
-                "Grand",
-                "Large",
-                "Large sans safe",
-                "Ligne",
-                "Nombreuses tours centrales",
-                "Tours inaccessibles",
-                "Plus"
-        );
-        labyrinthComboBox.setValue("Large");
-        labyrinthComboBox.setId("labyrinthComboBox");
-        labyrinthComboBox.setStyle("-fx-background-color: #f8f9fa; "
-                + "-fx-background-radius: 5; "
-                + "-fx-border-color: #dee2e6; "
-                + "-fx-border-radius: 5; "
-                + "-fx-padding: 5;");
-
-        VBox labyrinthSection = createSection("Choisir le labyrinthe", labyrinthComboBox);
-
-        // Section configuration des ennemis
-        TextField enemiesField = createStyledTextField("30", "enemiesField");
-        VBox enemiesSection = createSection("Nombre d'ennemis", enemiesField);
-
-        TextField nbEnnemiesToWinField = createStyledTextField("28", "ennemiesToWinField");
-        VBox objectifSection = createSection("Objectif d'ennemis à l'arrivée", nbEnnemiesToWinField);
-
-        // Section nombre de manches
-        TextField roundsField = createStyledTextField("20", "roundsField");
-        VBox roundsSection = createSection("Nombre de manches", roundsField);
-
-        // Section A*
-        CheckBox avecAstarBox = new CheckBox();
-        avecAstarBox.setSelected(true);
-        avecAstarBox.setId("avecOuSansAstarBox");
-        avecAstarBox.setStyle("-fx-padding: 5;");
-
-        VBox astarSection = createSection("Utiliser l'algorithme A*", avecAstarBox);
-
-        // Bouton de démarrage
-        Button startButton = new Button("Démarrer la partie");
-        startButton.setStyle("-fx-background-color: #4CAF50; "
-                + "-fx-text-fill: white; "
-                + "-fx-font-size: 14px; "
-                + "-fx-font-weight: bold; "
-                + "-fx-padding: 12 24; "
-                + "-fx-background-radius: 5; "
-                + "-fx-cursor: hand;");
-        startButton.setAlignment(Pos.CENTER);
-        startButton.setMaxWidth(Double.MAX_VALUE);
-
-        // Effet hover pour le bouton
-        startButton.setOnMouseEntered(e ->
-                startButton.setStyle("-fx-background-color: #45a049; "
-                        + "-fx-text-fill: white; "
-                        + "-fx-font-size: 14px; "
-                        + "-fx-font-weight: bold; "
-                        + "-fx-padding: 12 24; "
-                        + "-fx-background-radius: 5; "
-                        + "-fx-cursor: hand; "
-                        + "-fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.2), 5, 0, 0, 0);")
-        );
-
-        startButton.setOnMouseExited(e ->
-                startButton.setStyle("-fx-background-color: #4CAF50; "
-                        + "-fx-text-fill: white; "
-                        + "-fx-font-size: 14px; "
-                        + "-fx-font-weight: bold; "
-                        + "-fx-padding: 12 24; "
-                        + "-fx-background-radius: 5; "
-                        + "-fx-cursor: hand;")
-        );
-
-        configPrefaiteComboBox.valueProperty().addListener((observable, oldValue, newValue) -> {
-            if (newValue == "Aucune") {
-                // Réactiver les autres ComboBox pour permettre la modification
-                labyrinthComboBox.setDisable(false);
-                enemiesField.setDisable(false);
-                nbEnnemiesToWinField.setDisable(false);
-                roundsField.setDisable(false);
-                avecAstarBox.setDisable(false);
-            } else {
-                labyrinthComboBox.setDisable(true);
-                enemiesField.setDisable(true);
-                nbEnnemiesToWinField.setDisable(true);
-                roundsField.setDisable(true);
-                avecAstarBox.setDisable(true);
-                switch (newValue) {
-                    case "Chemin safe":
-                        labyrinthComboBox.setValue("Large");
-                        enemiesField.setText("15");
-                        nbEnnemiesToWinField.setText("12");
-                        roundsField.setText("10");
-                        avecAstarBox.setSelected(true);
-                        break;
-                    case "Aucun chemin safe":
-                        labyrinthComboBox.setValue("Large sans safe");
-                        enemiesField.setText("50");
-                        nbEnnemiesToWinField.setText("35");
-                        roundsField.setText("20");
-                        avecAstarBox.setSelected(true);
-                        break;
-                    case "Une tour unique":
-                        labyrinthComboBox.setValue("Ligne");
-                        enemiesField.setText("5");
-                        nbEnnemiesToWinField.setText("5");
-                        roundsField.setText("40");
-                        avecAstarBox.setSelected(true);
-                        break;
-                    case "Nombreuses tours centrales":
-                        labyrinthComboBox.setValue("Nombreuses tours centrales");
-                        enemiesField.setText("20");
-                        nbEnnemiesToWinField.setText("15");
-                        roundsField.setText("10");
-                        avecAstarBox.setSelected(true);
-                        break;
-                    case "Tours inaccessibles":
-                        labyrinthComboBox.setValue("Tours inaccessibles");
-                        enemiesField.setText("20");
-                        nbEnnemiesToWinField.setText("15");
-                        roundsField.setText("10");
-                        avecAstarBox.setSelected(true);
-                        break;
-                    default:
-                        break;
-                }
-            }
-        });
-
-        startButton.setOnMouseClicked(event -> {
-            laby.setStartTime();
-            String labyrinthString;
-            switch (labyrinthComboBox.getValue()) {
-                case "Plus":
-                    labyrinthString = openLaby();
-                    break;
-                default:
-                    labyrinthString = labyrinthMap.get(labyrinthComboBox.getValue());
-                    break;
-            }
-            dialogStage.close();
-            try {
-                laby.setUseAstar(avecAstarBox.isSelected());
-                ArrayList<Ennemy> ennemies = ModeleLabyrinth.createEnnemies(
-                        Integer.parseInt(enemiesField.getText())
-                );
-                laby.creerLabyrinthe(
-                        labyrinthString,
-                        ennemies,
-                        Integer.parseInt(roundsField.getText()),
-                        Integer.parseInt(nbEnnemiesToWinField.getText())
-                );
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-            startJeu(primaryStage);
-            MoteurJeu.labyFile = labyrinthString;
-        });
-
-        // Assemblage de l'interface
-        HBox sectionsContainer = new HBox(20);
-        VBox leftColumn = new VBox(20, configPrefaiteSection, enemiesSection, roundsSection);
-        VBox rightColumn = new VBox(20, labyrinthSection, objectifSection, astarSection);
-        sectionsContainer.getChildren().addAll(leftColumn, rightColumn);
-
-        root.getChildren().addAll(
-                titleLabel,
-                sectionsContainer,
-                startButton
-        );
-
-        // Configuration de la scène
-        Scene dialogScene = new Scene(root, 560, 525);
-        dialogScene.setFill(Color.TRANSPARENT);
-        dialogStage.setScene(dialogScene);
-        dialogStage.initOwner(primaryStage);
-        dialogStage.initModality(Modality.APPLICATION_MODAL);
-        Image icon = new Image("icon.png");
-        dialogStage.getIcons().add(icon);
-        dialogStage.showAndWait();
+        configWindow.show(primaryStage);
     }
 
-    // Méthode utilitaire pour créer une section
-    private VBox createSection(String labelText, Node content) {
-        VBox section = new VBox(10);
-        section.setStyle("-fx-background-color: white; "
-                + "-fx-padding: 15; "
-                + "-fx-background-radius: 5; "
-                + "-fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.05), 5, 0, 0, 0);");
-
-        Label label = new Label(labelText);
-        label.setStyle("-fx-font-size: 14px; "
-                + "-fx-font-weight: bold; "
-                + "-fx-text-fill: #34495e;");
-
-        section.getChildren().addAll(label, content);
-        return section;
-    }
-
-    // Méthode utilitaire pour créer un TextField stylisé
-    private TextField createStyledTextField(String defaultValue, String id) {
-        TextField field = new TextField(defaultValue);
-        field.setId(id);
-        field.setStyle("-fx-background-color: #f8f9fa; "
-                + "-fx-background-radius: 5; "
-                + "-fx-border-color: #dee2e6; "
-                + "-fx-border-radius: 5; "
-                + "-fx-padding: 8; "
-                + "-fx-font-size: 13px;");
-
-        // Effet focus
-        field.focusedProperty().addListener((obs, wasFocused, isFocused) -> {
-            if (isFocused) {
-                field.setStyle("-fx-background-color: white; "
-                        + "-fx-background-radius: 5; "
-                        + "-fx-border-color: #4CAF50; "
-                        + "-fx-border-radius: 5; "
-                        + "-fx-padding: 8; "
-                        + "-fx-font-size: 13px; "
-                        + "-fx-effect: dropshadow(three-pass-box, rgba(76,175,80,0.2), 5, 0, 0, 0);");
-            } else {
-                field.setStyle("-fx-background-color: #f8f9fa; "
-                        + "-fx-background-radius: 5; "
-                        + "-fx-border-color: #dee2e6; "
-                        + "-fx-border-radius: 5; "
-                        + "-fx-padding: 8; "
-                        + "-fx-font-size: 13px;");
-            }
-        });
-
-        return field;
-    }
 
     public void startJeu(Stage primaryStage) {
-        canvas.setWidth((getScreenSize().width / 7.0) * 6);
         canvas.widthProperty().bind(canvasContainer.widthProperty());
         canvas.heightProperty().bind(canvasContainer.heightProperty());
         SimpleMode mode = new SimpleMode();
@@ -493,25 +212,6 @@ public class MoteurJeu extends Application {
     }
 
 
-    private String openLaby() {
-        FileChooser fileChooser = new FileChooser();
-        fileChooser.setTitle("Sélectionner un fichier de labyrinthe");
-        FileChooser.ExtensionFilter extFilter =
-                new FileChooser.ExtensionFilter("Fichiers texte (*.txt)", "*.txt");
-        fileChooser.getExtensionFilters().add(extFilter);
-        fileChooser.setInitialDirectory(new File("Ressources"));
-        File selectedFile = fileChooser.showOpenDialog(null);
-
-        if (selectedFile != null) {
-            MoteurJeu.labyFile = selectedFile.getAbsolutePath();
-            return selectedFile.getAbsolutePath();
-        } else {
-            System.err.println("Aucun fichier sélectionné, utilisation du labyrinthe par défaut");
-            MoteurJeu.labyFile = "Ressources/Labyrinthe1.txt";
-            return "Ressources/Labyrinthe1.txt";
-        }
-    }
-
     private void openHelpWindow() {
         HelpWindow helpWindow = HelpWindow.getHelpWindow();
         helpWindow.show();
@@ -550,12 +250,8 @@ public class MoteurJeu extends Application {
                         "-fx-background-radius: 10px;"
         );
         restartButton.setOnAction(e -> {
-            try {
-                // On ferme la fenêtre actuelle
-                restartApp();
-            } catch (IOException ex) {
-                throw new RuntimeException(ex);
-            }
+            // On ferme la fenêtre actuelle
+            restartApp();
         });
 
         // Texte explicatif
@@ -589,34 +285,62 @@ public class MoteurJeu extends Application {
         stage.show();
     }
 
-    public void restartApp() throws IOException {
-        // Fermer l'application actuelle
-        Platform.exit();
-
-        // Relancer l'application en utilisant un nouveau processus Java
-        new Thread(() -> {
+    public void restartApp() {
+        Platform.runLater(() -> {
             try {
-                // Attendre que l'application actuelle se ferme complètement
-                Thread.sleep(2000);
+                // Fermer l'ancienne fenêtre
+                primaryStage.close();
 
-                // Construire la commande pour relancer l'application
-                String javaBin = System.getProperty("java.home") + File.separator + "bin" + File.separator + "java";
-                File currentJar = new File(getClass().getProtectionDomain().getCodeSource().getLocation().toURI());
+                // Créer une nouvelle instance de Stage
+                Stage newStage = new Stage();
+                primaryStage = newStage;
 
-                // Si le fichier actuel est un jar exécutable
-                if (currentJar.getName().endsWith(".jar")) {
-                    ArrayList<String> command = new ArrayList<>();
-                    command.add(javaBin);
-                    command.add("-jar");
-                    command.add(currentJar.getPath());
+                // Créer un nouveau modèle de labyrinthe
+                ModeleLabyrinth newLaby = new ModeleLabyrinth();
+                this.laby = newLaby;
 
-                    ProcessBuilder builder = new ProcessBuilder(command);
-                    builder.start();
-                }
+                // Mettre à jour l'instance du moteur avec le nouveau labyrinthe
+                MoteurJeu.setLaby(newLaby);
+
+                // Créer un nouveau canvas
+                Canvas newCanvas = new Canvas();
+                newCanvas.setWidth((getScreenSize().width / 7.0) * 6);
+                newCanvas.widthProperty().bind(canvasContainer.widthProperty());
+                newCanvas.heightProperty().bind(canvasContainer.heightProperty());
+
+                // Créer une nouvelle vue de labyrinthe pour le nouveau modèle
+                ViewLabyrinth newViewLabyrinth = new ViewLabyrinth(newLaby, newCanvas);
+                newLaby.registerObserver(newViewLabyrinth);
+
+                // Ajouter le nouveau Canvas au Pane
+                canvasContainer.getChildren().clear();  // Supprimer l'ancien Canvas s'il existe
+                canvasContainer.getChildren().add(newCanvas);  // Ajouter le nouveau Canvas
+
+                // Réinitialiser FPS
+                FPS = BASE_FPS;
+
+                // Créer une nouvelle fenêtre de configuration avec le labyrinthe réinitialisé
+                ConfigWindow newConfigWindow = new ConfigWindow(newLaby, newStage, newCanvas, this);
+
+                // Enregistrer les nouvelles références
+                this.canvas = newCanvas;
+                this.configWindow = newConfigWindow;
+
+                // Réinitialiser le temps et les statistiques du jeu
+                newLaby.resetStats();
+
+                // Ajouter l'icône à la nouvelle fenêtre
+                Image icon = new Image("icon.png");
+                newStage.getIcons().add(icon);
+                newStage.setTitle("Tower Defense");
+
+                // Afficher la nouvelle fenêtre de configuration
+                newConfigWindow.show(newStage);
+
             } catch (Exception e) {
                 e.printStackTrace();
             }
-        }).start();
+        });
     }
 
 
@@ -638,3 +362,4 @@ public class MoteurJeu extends Application {
         HBox.setMargin(objectifLabel, new Insets(0, 0, 0, 20));
     }
 }
+
